pyproject.toml
[project]
name = "pr-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.168.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp[cli]>=1.6.0",
    "notion-client>=2.3.0",
    "pydrive>=1.3.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
